Fix issues on serialization/deserizaliztion from elasticsearch + synchronisation implementation : round 2

+++ CanopeeAgent/CanopeeServer/Datas/Entities/AgentGroup.cs
@@ -14,7 +14,7 @@ namespace CanopeeServer.Datas.Entities
 
         public string Group { get; set; }
         
-        public int Priority { get; set; }
+        public long Priority { get; set; }
 
         public DateTime EventDate { get; set; }
